Remove component update command
