+++ src/main/java/com/aws/greengrass/cli/commands/ComponentCommand.java
@@ -9,7 +9,6 @@ import com.aws.greengrass.cli.adapter.NucleusAdapterIpc;
 import com.aws.greengrass.ipc.services.cli.exceptions.CliIpcClientException;
 import com.aws.greengrass.ipc.services.cli.exceptions.GenericCliIpcServerException;
 import com.aws.greengrass.ipc.services.cli.models.ComponentDetails;
-import com.aws.greengrass.ipc.services.cli.models.CreateLocalDeploymentRequest;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import picocli.CommandLine;
@@ -51,41 +50,6 @@ public class ComponentCommand extends BaseCommand {
         for (String componentName : componentNames) {
             nucleusAdapterIpc.stopComponent(componentName);
         }
-        return 0;
-    }
-
-
-    // GG_NEEDS_REVIEW: TODO: deprecate this for "create" sub command under deployment command space. (pending UAT update)
-    // GG_NEEDS_REVIEW: TODO: input validation and better error handling https://sim.amazon.com/issues/P39478724
-    @CommandLine.Command(name = "update",
-            description = "Updates Greengrass applications with provided recipes, artifacts, and runtime parameters")
-    public int deploy
-    (@CommandLine.Option(names = {"-m", "--merge"}, paramLabel = "Component and version") Map<String, String> componentsToMerge,
-     @CommandLine.Option(names = {"--remove"}, paramLabel = "Component Names") List<String> componentsToRemove,
-     @CommandLine.Option(names = {"-g", "--groupId"}, paramLabel = "group Id") String groupId,
-     @CommandLine.Option(names = {"-r", "--recipeDir"}, paramLabel = "Recipe Folder Path") String recipeDir,
-     @CommandLine.Option(names = {"-a", "--artifactDir"}, paramLabel = "Artifacts Folder Path") String artifactDir,
-     @CommandLine.Option(names = {"-p", "--param"}, paramLabel = "Runtime parameters") Map<String, String> parameters,
-     @CommandLine.Option(names = {"-c", "--update-config"}, paramLabel = "Update configuration") String configUpdate)
-            throws CliIpcClientException, GenericCliIpcServerException, JsonProcessingException {
-        // GG_NEEDS_REVIEW: TODO Validate folder exists and folder structure
-        Map<String, Map<String, Object>> componentNameToConfig = convertParameters(parameters);
-        Map<String, Map<String, Object>> configurationUpdate = null;
-        if (configUpdate != null && !configUpdate.isEmpty()) {
-            configurationUpdate = mapper.readValue(configUpdate, Map.class);
-        }
-        if (recipeDir != null || artifactDir != null) {
-            nucleusAdapterIpc.updateRecipesAndArtifacts(recipeDir, artifactDir);
-        }
-        CreateLocalDeploymentRequest createLocalDeploymentRequest = CreateLocalDeploymentRequest.builder()
-                .groupName(groupId)
-                .configurationUpdate(configurationUpdate)
-                .componentToConfiguration(componentNameToConfig)
-                .rootComponentVersionsToAdd(componentsToMerge)
-                .rootComponentsToRemove(componentsToRemove)
-                .build();
-        String deploymentId = nucleusAdapterIpc.createLocalDeployment(createLocalDeploymentRequest);
-        System.out.println("Local deployment has been submitted! Deployment Id: "+ deploymentId);
         return 0;
     }
